Normalize input columns when loading datasets.
When we append a dataframe to the input dataset frame, make sure we add any
column that have been declared in the DataSetContainer::fields list but is not
found in the input dataframe. This ensures that any operation performed later
can safely assume that all declared columns are present (even if filled with a
default value).