Rework the Target descriptors.
Targets that identify files now can either refer to a single path or a
group of paths that represent different iterations of the same
benchmark.
The Task.ouptuts() method now yields a pair of (target-name,
target), the target name is meant to be used by other tasks that need
to reference a specific output target from another task.