Refactor flattened struct layout construction.
This now also uses a single transaction for the whole insertion.
Change the primary key for the member_bounds table to use (owner,
name), this ensures that conflicting names do not violate the
uniqueness constraint. This may occur if a structure with the same
name is present in different scopes.