Ditch the old benchmark_matrix system.
The session now uses a better parameterization approach. This is a
more standard polars dataframe layout. There are no special index
columns, instead we have normal columns for user-defined
parameterization axes, an implicit column for the 'instance'
parameterization axis and a reserved 'descriptor' column that contains
the Benchmark instance corresponding to the row.