Start deprecation of the old instance configuration.
Now we have a system configuration embedded in the benchmark_config.
The system configuration associates instance configurations to benchmark
contexts based on parameter combinations, this translates to the 'target'
parameterization column by default, although a custom target column may
be specified.

The benchmark config is now unique, because the ability to have multiple
benchmarks was useless and complicated.