First cut at implementing arbitrary parameter axes.
This adds two base analysis tasks, the GenericAnalysisTask and
the SliceAnalysisTask.
The GenericAnalysisTask is a root task that should be added to the
analysis configuration. It is used to schedule a number of SliceAnalysisTasks
that are responsible for the analysis of a slice of the data, across an
arbitrary set of "fixed" parameterization axes.

This means that there will be a SliceAnalysisTask for each combination of
the "fixed" axes. Each instance of SliceAnalysisTask receives a different
chunk of data.

This is currently demonstrated by the generic execution and
generic timing plot tasks.

There are limitations to how this is represented in the analysis configuration
at it is slightly more verbose than necessary.