coords: Improve check for dataframe misalignment.
The plot coordinate allocator requires the dataframe to be aligned
along the shift_by and stack_by axes.
If this is not satisfied, it is not possible to generate the coordinates in the
current implementation.
This is considered an error in the caller, which has the responsibility to
ensure a sensible data shape.
We could realign the dataframe but this requires knowing the parameter columns
and any extra index colum generated in the analysis task.