plot: Rework plot grid configuration.
This introduces a new system for column transformations in order to
better support last-stage transformations on the data for
visualisation.

This adds a new derived_columns configuration field that allows to
specify column transformations to apply to the dataframe before
plotting.
The grid configuration now changes to use ColRef names in all places
where a column name is expected. A ColRef is a bracketed identifier
such as <my_ref>, which uniquely identifies a dataframe column. This
is useful because the name is stable and can be reused in the
configuration multiple times without updating it if the name of the
column is changed for presentation purposes.

Simplify the grid label configuration. The grid has split
responsability now:
 - the plot implementations are expected to set X, Y labels as usual.
 - the grid implementation suppresses unwanted labels depending on
 axis sharing configuration and overrides the labels if tile_xlabel or
 tile_ylabel are configured.